+++ data/web/sensorconfig.js
@@ -16,6 +16,7 @@ function GetInitData() {
 function MyCallback() {
   CreateSelectionListFromInputField('input[type=number][id^=GpioPin]', [gpio]);
   CreateSelectionListFromInputField('input[type=number][id^=AnalogPin]', [gpioanalog]);
+  validate_identifiers("moistureRows");
 }
 
 // ************************************************