+++ src/sensor.cpp
@@ -12,7 +12,8 @@ sensor::sensor() :
   ads1115_i2c(48),
   ads1115_port(0),
   threshold_min(26),
-  threshold_max(30) {
+  threshold_max(30),
+  moistureEnabled(false) {
   
   #ifdef ESP8266
     uint8_t pinAnalogDefault = 0;
@@ -20,6 +21,10 @@ sensor::sensor() :
     uint8_t pinAnalogDefault = 36; // ADC1_CH0 (GPIO 36) 
   #endif
   
+  #ifdef USE_ADS1115
+    ads1115_devices = new std::vector<adsdev_t>{};
+  #endif
+
   this->pinAnalog = pinAnalogDefault;
   
   LoadJsonConfig(); 
@@ -106,75 +111,153 @@ void sensor::loop_hcsr04() {
 }
 
 #ifdef USE_ADS1115
-  void sensor::init_ads1115(uint8_t i2c, uint8_t port) {
-    if (Config->GetDebugLevel() >=4) Serial.printf("Init ADS1115 at i2cAdress 0x%02x \n", i2c);
+  void sensor::init_ads1115(uint8_t i2c, uint8_t port, String topic) {  
+    adsdev_t* device = this->getAdsDevice(i2c);
     
-    this->ads1115_i2c = i2c;
-    this->ads1115_port = port;
-    setSensorType(ADS1115);
-    
-    ADS1115_WE* adc = new ADS1115_WE(0x48);
-  
-    if(!adc->init()){
-      Serial.printf("Could not connect to ADS1115 at i2cAdress 0x%02x \n", i2c );
+    if (device == nullptr) { 
+      if (Config->GetDebugLevel() >=4) Serial.printf("Init ADS1115 at i2cAdress 0x%02x \n", i2c);
+      
+      adsdev_t ads;
+      ads.device = ADS1115_WE(i2c);
+      
+      if(!ads.device.init()){
+        if (Config->GetDebugLevel() >=1) Serial.printf("Could not connect to ADS1115 at i2cAddress 0x%02x, ignore it!\n", i2c );
+      } else {
+        if (Config->GetDebugLevel() >=3) Serial.printf("Initialize ADS1115 at i2cAddress 0x%02x with channel %d\n", i2c, port);
+        ads.device.setVoltageRange_mV(ADS1115_RANGE_4096);
+        ads.i2cAddress = i2c;
+        if (port == 0) {ads.topic_chan1 = topic;}
+        if (port == 1) {ads.topic_chan2 = topic;}
+        if (port == 2) {ads.topic_chan3 = topic;}
+        if (port == 3) {ads.topic_chan4 = topic;}
+        this->ads1115_devices->push_back(ads);
+      }
     } else {
-      adc->setVoltageRange_mV(ADS1115_RANGE_4096);
-      this->Device = adc;
+      if (Config->GetDebugLevel() >=3) Serial.printf("Add Channel %d to ADS1115 at i2cAddress 0x%02x with topic '%s' \n", port, i2c, topic);
+      if (port == 0) {device->topic_chan1 = topic;}
+      if (port == 1) {device->topic_chan2 = topic;}
+      if (port == 2) {device->topic_chan3 = topic;}
+      if (port == 3) {device->topic_chan4 = topic;}
     }
   }
 
-  void sensor::loop_ads1115() {
+  
+  /*returns the proper ADS1115 vector of given i2c-address */
+  adsdev_t* sensor::getAdsDevice(uint8_t i2c) {
+    for (uint8_t i=0; i<this->ads1115_devices->size(); i++) {
+      if (this->ads1115_devices->at(i).i2cAddress == i2c) {
+        return &this->ads1115_devices->at(i);
+      }
+    }
+    return nullptr;
+  }
+
+  /*returns the proper ADS1115_MUX channel of given port number*/
+  ADS1115_MUX sensor::getAdsChannel(uint8_t port) {
+    switch (port) {
+      case 0:
+        return ADS1115_COMP_0_GND;
+        break;
+      case 1:
+        return ADS1115_COMP_1_GND;
+        break;
+      case 2:
+        return ADS1115_COMP_2_GND;
+        break;
+      case 3:
+        return ADS1115_COMP_3_GND;
+        break;
+      default:
+        Serial.printf("ADS115 portnummer %d not available \n", port);
+        break;
+    }
+    return ADS1115_COMP_0_GND;
+  }
+
+  void sensor::loop_ads1115_moisture() {
+    uint16_t raw = 0;
+    uint8_t  level = 0;
+    
+    for (uint8_t i=0; i<this->ads1115_devices->size(); i++) {
+      for (uint8_t chan=0; chan<4; chan++) {
+        if (this->ads1115_devices->at(i).i2cAddress == this->ads1115_i2c && chan == this->ads1115_port) {
+          // exclude sensor channel from measurement the moisture
+          continue;
+        }  
+        if (chan == 0 && this->ads1115_devices->at(i).topic_chan1.length()==0) continue;  
+        if (chan == 1 && this->ads1115_devices->at(i).topic_chan2.length()==0) continue;
+        if (chan == 2 && this->ads1115_devices->at(i).topic_chan3.length()==0) continue;
+        if (chan == 3 && this->ads1115_devices->at(i).topic_chan4.length()==0) continue;
+
+        raw = readADS1115Channel(&this->ads1115_devices->at(i), this->getAdsChannel(chan));
+        level = map(raw, 0, 3300, 0, 100); // 0-3.3V -> 0-100%
+        String topic = "";
+        switch (chan) {
+          case 0:
+            topic = this->ads1115_devices->at(i).topic_chan1;
+            break;
+          case 1:
+            topic = this->ads1115_devices->at(i).topic_chan2;
+            break;
+          case 2:
+            topic = this->ads1115_devices->at(i).topic_chan3;
+            break;
+          case 3:
+            topic = this->ads1115_devices->at(i).topic_chan4;
+            break;
+        } 
+
+        if (raw > 0 ) { mqtt->Publish_Int(topic.c_str(), (int)level, false); }
+      }
+    }
+  }
+
+  void sensor::loop_ads1115_sensor() {
     this->raw = 0;
     this->level = 0;
 
-    if (!this->Device) {
+    if (!this->getAdsDevice(this->ads1115_i2c)) {
       if (Config->GetDebugLevel() >=3) Serial.printf("Measure of analog Sensor ADS1115 port %d requested, but not ADS1115 found. Stop measure! \n", this->ads1115_port);
     } else {
 
       if (Config->GetDebugLevel() >=4) Serial.printf("start measure, use analog Sensor ADS1115 port: %d \n", this->ads1115_port);
     
-      switch (this->ads1115_port) {
-        case 0:
-          this->raw = readADS1115Channel(ADS1115_COMP_0_GND);
-        break;
-        case 1:
-          this->raw = readADS1115Channel(ADS1115_COMP_1_GND);
-        break;
-        case 2:
-          this->raw = readADS1115Channel(ADS1115_COMP_2_GND);
-        break;
-        case 3:
-          this->raw = readADS1115Channel(ADS1115_COMP_3_GND);
-        break;
-        default:
-          Serial.printf("portnummer %d not available \n", this->ads1115_port);
-        break;
-      }
+      this->raw = readADS1115Channel(this->getAdsDevice(this->ads1115_i2c), this->getAdsChannel(this->ads1115_port));
+      
       this->level = map(this->raw, measureDistMin, measureDistMax, 0, 100); // 0-100%
     }
   }
 
-  uint16_t sensor::readADS1115Channel(ADS1115_MUX channel) {
+  uint16_t sensor::readADS1115Channel(adsdev_t* device, ADS1115_MUX channel) {
     int16_t raw = 0;
-    ADS1115_WE* adc = static_cast<ADS1115_WE*>(this->Device);
-    adc->setCompareChannels(channel);
-    adc->startSingleMeasurement();
-    while(adc->isBusy()){}
-    raw = adc->getResultWithRange(-4096,4096); 
+    
+    device->device.setCompareChannels(channel);
+    device->device.startSingleMeasurement();
+    while(device->device.isBusy()){}
+    raw = device->device.getResultWithRange(-4096,4096); 
     return (uint16_t) abs(raw);
   }
 #endif
 
 void sensor::loop() {
-  if (millis() - this->previousMillis > this->measurecycle*1000) {
-    this->previousMillis = millis();
+  /*start measuring soil moisture, every 60sec */
+#ifdef USE_ADS1115
+  if (millis() - this->previousMillis_moisture > 60*1000) {
+    this->previousMillis_moisture = millis();
+    loop_ads1115_moisture();
+  }
+#endif
+
+  /*start measuring sensor*/
+  if (millis() - this->previousMillis_sensor > this->measurecycle*1000) {
+    this->previousMillis_sensor = millis();
    
     if (this->Type == ONBOARD_ANALOG) {loop_analog();}
    
     if (this->Type == HCSR04) {loop_hcsr04();}
 
     #ifdef USE_ADS1115
-      if (this->Type == ADS1115) {loop_ads1115();}
+      if (this->Type == ADS1115) {loop_ads1115_sensor();}
     #endif
 
     if (this->Type != NONE && this->level !=0 && Config->Enabled3Wege()) {
@@ -200,6 +283,8 @@ void sensor::loop() {
 
 void sensor::LoadJsonConfig() {
   mqtt->ClearSubscriptions(MyMQTT::SENSOR);
+  this->ads1115_devices->clear();
+
   String selection = "";
 
   if (LittleFS.exists("/sensorconfig.json")) {
@@ -235,16 +320,23 @@ void sensor::LoadJsonConfig() {
           if (elem.containsKey("ads1115_port"))         { this->ads1115_port = elem["ads1115_port"].as<int>();}
           if (elem.containsKey("externalSensor"))       { this->externalSensor = elem["externalSensor"].as<String>();}
           if (elem.containsKey("selection"))            { selection = elem["selection"].as<String>(); }
+          if (elem.containsKey("sel_moisture"))         { if (elem["sel_moisture"].as<String>() == "on") {this->moistureEnabled = true;} else {this->moistureEnabled = false;}}
+
+          if (elem.containsKey("mqtttopic") && 
+              elem.containsKey("ads_addr") &&
+              elem.containsKey("ads_port")) {
+                this->init_ads1115(strtoul(elem["ads_addr"].as<String>().c_str(), NULL, 16), elem["ads_port"].as<int>(), elem["mqtttopic"].as<String>());
+              }
         }
       } while (stream.findUntil(",","]"));
 
-      if (selection == "analog")        { init_analog(this->pinAnalog); }
-      else if (selection == "hcsr04")   { init_hcsr04(this->pinTrigger, this->pinEcho); }
-      else if (selection == "extern")   { init_extern(this->externalSensor); }
-      else if (selection == "none")     { this->Type=NONE; Serial.println(F("No LevelSensor requested")); } 
+      if (selection == "analog")        { this->init_analog(this->pinAnalog); }
+      else if (selection == "hcsr04")   { this->init_hcsr04(this->pinTrigger, this->pinEcho); }
+      else if (selection == "extern")   { this->init_extern(this->externalSensor); }
+      else if (selection == "none")     { this->setSensorType(NONE); Serial.println(F("No LevelSensor requested")); } 
             
       #ifdef USE_ADS1115  
-        else if(selection == "ads1115") { init_ads1115(this->ads1115_i2c, this->ads1115_port); }
+        else if(selection == "ads1115") { this->setSensorType(ADS1115); this->init_ads1115(this->ads1115_i2c, this->ads1115_port); }
       #endif
 
     } else {
@@ -271,10 +363,72 @@ void sensor::GetInitData(AsyncResponseStream *response) {
   json["data"]["sel3"] = ((this->Type==ADS1115)?1:0);
   json["data"]["ads1115_i2c"] = ads1115_i2c_hex.str(); 
   json["data"]["ads1115_port"] = this->ads1115_port;
+  json["data"]["sel_moisture1"] = ((this->moistureEnabled)?0:1);
+  json["data"]["sel_moisture2"] = ((this->moistureEnabled)?1:0);
+  
+  json["data"]["rows"].to<JsonArray>();
+  
+  uint8_t counter = 0;
+  for (uint8_t i=0;i<this->ads1115_devices->size(); i++) {
+    std::ostringstream moisture_i2c_hex;
+    moisture_i2c_hex << std::hex << std::uppercase << std::setw(2) << std::setfill('0') << (int)this->ads1115_devices->at(i).i2cAddress;
+    if (this->ads1115_devices->at(i).topic_chan1.length() > 0 &&
+        !(this->GetType() == ADS1115  &&
+         this->ads1115_i2c == this->ads1115_devices->at(i).i2cAddress &&
+         this->ads1115_port == 0
+        )) {
+      json["data"]["rows"][counter]["mqtttopic"] = this->ads1115_devices->at(i).topic_chan1.c_str();
+      json["data"]["rows"][counter]["ads_port"] = 0;
+      json["data"]["rows"][counter]["ads_addr"] = moisture_i2c_hex.str(); 
+      counter++;
+    }
+
+    if (this->ads1115_devices->at(i).topic_chan2.length() > 0 &&
+        !(this->GetType() == ADS1115  &&
+         this->ads1115_i2c == this->ads1115_devices->at(i).i2cAddress &&
+         this->ads1115_port == 0
+        )) {
+      json["data"]["rows"][counter]["mqtttopic"] = this->ads1115_devices->at(i).topic_chan2.c_str();
+      json["data"]["rows"][counter]["ads_port"] = 1;
+      json["data"]["rows"][counter]["ads_addr"] = moisture_i2c_hex.str(); 
+      counter++;
+    }
+
+    if (this->ads1115_devices->at(i).topic_chan3.length() > 0 &&
+        !(this->GetType() == ADS1115  &&
+         this->ads1115_i2c == this->ads1115_devices->at(i).i2cAddress &&
+         this->ads1115_port == 0
+        )) {
+      json["data"]["rows"][counter]["mqtttopic"] = this->ads1115_devices->at(i).topic_chan3.c_str();
+      json["data"]["rows"][counter]["ads_port"] = 2;
+      json["data"]["rows"][counter]["ads_addr"] = moisture_i2c_hex.str(); 
+      counter++;
+    }
+
+    if (this->ads1115_devices->at(i).topic_chan4.length() > 0 &&
+        !(this->GetType() == ADS1115  &&
+         this->ads1115_i2c == this->ads1115_devices->at(i).i2cAddress &&
+         this->ads1115_port == 0
+        )) {
+      json["data"]["rows"][counter]["mqtttopic"] = this->ads1115_devices->at(i).topic_chan4.c_str();
+      json["data"]["rows"][counter]["ads_port"] = 3;
+      json["data"]["rows"][counter]["ads_addr"] = moisture_i2c_hex.str(); 
+      counter++;
+    }
+  }
+
+  // print template as first row
+  if (counter == 0) {
+    json["data"]["rows"][0]["mqtttopic"] = "moisture_1";
+    json["data"]["rows"][0]["ads_port"] = 0;
+    json["data"]["rows"][0]["ads_addr"] = "48";
+  }
+
 #else 
   json["data"]["sel_ADS1115"]["className"] = "hide";
   json["data"]["ads1115_0"]["className"] = "hide";
   json["data"]["ads1115_1"]["className"] = "hide";
+  json["data"]["moistureTable"]["className"] = "hide";
 #endif
 
   json["data"]["sel4"] = ((this->Type==EXTERN)?1:0);